+++ web/src/components/Section.js
@@ -3,6 +3,7 @@ import React, { Fragment } from 'react';
 
 import Instruction from './Instruction';
 import { t } from '../i18n';
+import SecondaryNav from './SecondaryNav';
 
 const SectionTitle = ({ children }) => <h2>{children}</h2>;
 
@@ -25,6 +26,7 @@ const Section = ({ children, id, resource }) => {
       <h2 className="ds-h2">{title}</h2>
       <span className="ds-text--lead">{helptext}</span>
       {children}
+      <SecondaryNav/>
     </section>
   );
 };
