+++ web/src/containers/activity/Overview.js
@@ -15,6 +15,7 @@ import { Subsection } from '../../components/Section';
 import { NameAndFundingSourceForm } from './NameAndFundingSource';
 import TextArea from '../../components/TextArea';
 import { selectActivityByIndex } from '../../reducers/activities.selectors';
+import {SecondaryNav} from '../../components/SecondaryNav';
 
 const ActivityOverview = ({
   activity,
