+++ web/src/styles/index.scss
@@ -21,6 +21,7 @@
 @import '/scss/rte';
 @import '/scss/save-button';
 @import '/scss/section';
+@import '/scss/secondary-nav';
 
 .nowrap {
   white-space: nowrap;
